Begins to integrate the form functionality for the EphemeraBox and EphemeraFolder resources within EphemeraProjects
Adding routing and minimal support for EphemeraFolders

Implementing a first attempt at supporting EphemeraBox membership and improving the routes for EphemeraBoxes and EphemeraFolders

Ensuring that Resources can have IIIF Manifests exposed using the class method "can_have_manifests?"; Ensuring that EphemeraProjects and EphemeraBoxes are linked using the ContextualPath

Implementing the attributes for the EphemeraFolders; Resolving issues for ScannedResources

Integrating support for membership within EphemeraFolders

Tracking files being merge after the latest rebase

Adding the Factories and test suites for the Ephemera Decorators

Extending the test coverage for the Ephemera* Controllers and Decorators

Don't create a real property for append_id

Append_id is a virtual property in Valkyrie that exists on ChangeSets.
It's not something we want to persist - it's something we act on in a
ChangeSetPersister.

Extending the coverage for the Ephemera Decorators; Adding test coverage for the EphemeraBox and Folder ChangeSets

Increasing the test coverage for EphemeraFolders and the ephemera editor role

Clean up box show view.

Ensures that IIIF Manifests are properly structured for EphemeraFolders

Add Ephemera Folder Workflow

Also DRYs this up a bit to make it easier to add another workflow later.

Implementing the EphemeraVocabulary Class; Customizing and extending the View templates for EphemeraFields and EphemeraVocabulary objects

Increasing the test coverage for Ephemera Vocabularies

Extending the coverage for the EphemeraController

Resolving a typo. error in the parent breadcrumbs partial

Linking EphemeraFolders to vocabulary terms for EphemeraFields

Ensures that errors encountered in EphemeraFields are handled properly

Extending the coverage for Ephemera* Controllers and access permissions