Recursively search for FileSets in Manifest ThumbnailBuilder.
When collection manifests have multi-volume works the thumbnail_id of
the MVW points to the scanned resource within it that has the thumbnail
it should use, rather than the FileSet. The thumbnail builder was
assuming it'd get back a FileSet, which caused an error to be thrown
while embedding thumbnails in those manifests. The error can be seen
here: https://app.honeybadger.io/projects/53391/faults/37473184

The change here iterates down thumbnail_ids until it finds a FileSet,
and if it doesn't it just doesn't append a thumbnail to the manifest.