Change the signature of DerivativeService to accept an ID
Work towards #2657

It turns out that if DerivativeServices are going to save, there's a
good chance that they're going to step over one another. Instead, they
should pull the resource at the beginning of their work and save at the
end, so we can enforce that by making the method signature take an ID
instead of a change_set.

While this is part of a band-aid solution, it would be necessary work in
order to make optimistic locking work too (the locks for each subsequent
save would be outdated otherwise.)

Co-authored-by: Anna Headley <anna3LC@gmail.com>