Stop pre-populating everywhere.
ChangeSet prepopulation is meant to prepare a ChangeSet for display in
an HTML form. It spins up things like nested form defaults so that they
can be rendered. Instead of that, we were hooking into it to convert
multi-valued fields into single-valued ones. By doing that at
initialization instead, we can start following Reform documentation.