FileSet tombstones should reference parent.
This also deletes child tombstones on deletion of the parent, with the
assumption that we'd just want to restore the whole resource at that
point. It also gets around the fact that the parent is gone when the
children start getting deleted, which would make it impossible to make
tombstones for them.