Refactor download permissions to act on the FileMetadata
There's no point in checking for the FileSet, as the permissions for
streaming differ from the permissions for downloading the original file.
Instead, derivatives are allowed if the FileSet is readable (it's
complete and accessible by the user), and original files are downloadable if the
FileSet is configured to be so.

An exception is made for Geo Thumbnails/Metadata, which are set to
always be downloadable as per existing functionality.

Closes #3652