PDF Derivative Service should extract/move files in batches.
The lazy array makes it extract groups of 200 files at a time, then adds
them to the parent (in a transaction), which moves the files off the tmp
location. This should take care of the issue where we run out of space,
but if the extraction fails we may end up with orphan files in
production.