Test FacetIndexer from Solr directly
The tests for this class actually hit the class twice, once when the
record is created (and the metadata is pulled), and then again when
`described_class.to_solr` is called. The first time dates come through
as Time objects. The second time they come through as strings. There was
a bug in the code giving a nil value when the date was a Time object,
which the tests didn't catch because they were only checking values via
`described_class`. Adding a test that looks at the index directly gave a
failing test from which to resolve the bug.