BibID Updater Fixes
I found that the cursor used in the find_by_property query forces a
transaction, meaning any changes weren't reflected in the edit forms of
the UI, and resulted in Solr being in a different state than postgres.

There's plenty of memory to load it in, and it only takes a minute or
so.