Migrate specs to use alma IDs
In some cases the tests still need to use the older style voyager IDs,
and there we force the old / shorter ID into place, skipping validation.

Advances #5679 but this work is not complete. To find more IDs that need
to be migrated, update the `def catalog?` method in
PulMetadataServices::Client and run the suite.