Fix EZID type mismatch error.
When an HTTP ACCEPT header is sent which is a wildcard, Rails picks the
first response in a `respond_to` block. Blacklight runs first, which has
HTML first, and then it'd hit our catch block and do JSON first, which
would then error because Rails tries to respond with both HTML and JSON.

I couldn't get the spec to replicate this behavior (something's up with
the testing harness), but I was able to
make sure that this method does the HTML behavior when given a wildcard
header.

Closes #5350