Break up manifest_builder_v3 subclasses into separate files
This will make it easier to search within a file and know which class
you're looking at when you see results. Also clarifies the relationship
of `#manifest` to the main class, since it was at the very bottom after
all the subclasses, and reduces the length of the long file.

refs #5747